+++ app/src/main/java/com/forrestguice/suntimeswidget/calculator/SuntimesMoonData.java
@@ -19,7 +19,6 @@
 package com.forrestguice.suntimeswidget.calculator;
 
 import android.content.Context;
-import android.util.Log;
 
 import com.forrestguice.suntimeswidget.SuntimesUtils;